+++ Dawn Winery/Data/Receipes.cs
@@ -6,17 +6,17 @@ namespace Dawn_Winery.Data
     {
         public bool Type { get; set; }
         public string? Grape1 { get; set; }
-        public int G1Kilo { get; set; }
+        public float G1Kilo { get; set; }
         public string? Grape2 { get; set; }
-        public int? G2Kilo { get; set; }
+        public float? G2Kilo { get; set; }
         public string? Grape3 { get; set; }
-        public int? G3Kilo { get; set; }
+        public float? G3Kilo { get; set; }
         public string? Grape4 { get; set; }
-        public int? G4Kilo { get; set; }
+        public float? G4Kilo { get; set; }
         public string? Grape5 { get; set; }
-        public int? G5Kilo { get; set; }
+        public float? G5Kilo { get; set; }
         public string? Grape6 { get; set; }
-        public int? G6Kilo { get; set; }
+        public float? G6Kilo { get; set; }
         public int SO2 { get; set; }
         public string? Rname { get; set; }
