+++ Dawn Winery/Models/RawMaterial.cs
@@ -12,7 +12,7 @@ namespace Dawn_Winery.Models
         public int Acidity { get; set; }
         public int Body { get; set; }
         public int Tannin { get; set; }
-        public int Stock { get; set; }
+        public float Stock { get; set; }
         public int Quality { get; set; }
     }
 }
